// packages/headless/src/modules/paymentDetails/gateways/types.d.ts
export enum GatewayTypes {
  FREE = -1,
  STORED = 0,
  // ---
  CARD = 1,
  BANK_TRANSFER = 2,
  DIRECT_DEBIT = 3,
  SEPA = 4,
  OFFLINE = 5,
  MOBILE = 6,
  WALLET = 7,
}
// ...
export enum GatewayCtx {
  PAY = "pay", // PAY = Gateways are shown in the context of making a payment (invoice, topup etc)
  ADD = "add", // ADD = Gateways are shown in the context of adding a stored payment detail
}

export enum GatewayProviderCodes {
  STRIPE = "Stripe_PaymentIntents",
}